requirements.txt
Django>=5.1
djangorestframework>=3.15
psycopg2-binary>=2.9
lxml>=5.1
WeasyPrint>=62
openpyxl>=3.1
matplotlib>=3.8
passlib>=1.7
Pillow>=10.0
whitenoise>=6.5
openai>=1.0
